requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
defusedxml==0.7.1
idna==3.3
importlib-metadata==4.11.3
jira==3.2.0
keyring==23.5.0
mdutils==1.2.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.2
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
tcms-api==11.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
beautifulsoup4==4.11.1
python-gitlab==2.1.0
